// frontend/src/app/core/services/api/model/clubCoachCreateRequest.ts
/**
 * BFG API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Request payload for assigning a coach to a club
 */
export interface ClubCoachCreateRequest { 
    /**
     * UUID of the user to assign as coach
     */
    coachId: string;
    /**
     * UUID of the club to assign the coach to
     */
    clubId: string;
}